Migration to change OpenIdAssociation.server_url to a string rather than binary
When passing the server_url back to the database from
lib/masq/active_record_openid_store/openid_ar_store.rb some databases
(eg oracle) expect the string being passed back to be a blob - since its
a binary field - and blow up when its not.

Since server_url is a string column in other masq tables and its a url
so it doesnt need to be case sensitive (binary columns are case sensitive in mysql)
we can safely convert this column to a string.